+++ src/pcot/xforms/xformmanualregister.py
@@ -460,6 +460,9 @@ class TabManualReg(pcot.ui.tabs.Tab):
             self.mark()
             self.node.type.delSelPoint(self.node)
             self.changed()
+        else:
+            return False
+        return True
 
     def canvasMouseMoveEvent(self, x, y, e):
         if self.mouseDown: